Gave up on QScintilla
There are 13 separate steps involved in adding a new lexer to the main
library, and no real support for adding them externally as plugins.
While the editor itself is very feature-complete it would take too long
to add GP program support and then condition support for rule editing.
Instead a custom widget will be created to handle editing similar to
the CodeEditor example in the Qt4 examples.